Adjust spacing of static_cast<> calls to follow Octave coding conventions.
* __init_qt__.cc, file-editor.cc, main-window.cc, bsxfun.cc, data.cc,
filter.cc, gammainc.cc, ls-hdf5.cc, mex.cc, oct-stream.cc, lo-mappers.cc,
lo-specfun.cc, data-conv.cc, lo-regexp.cc, oct-inttypes.h:
Adjust spacing of static_cast<> calls to follow Octave coding conventions.

+++ libgui/graphics/__init_qt__.cc
@@ -199,7 +199,7 @@ DEFUN (__uigetfile_qt__, args, , "")
   else
     retval(0) = "";
   retval(1) = "";
-  retval(2) = static_cast<double> (0);
+  retval(2) = 0.0;
 
   if (defaultFileName.isEmpty ())
     defaultFileName = defaultDirectory;
@@ -277,7 +277,7 @@ DEFUN (__uiputfile_qt__, args, , "")
 
   retval(0) = "";
   retval(1) = "";
-  retval(2) = static_cast<double> (0);
+  retval(2) = 0.0;
 
   if (defaultFileName.isEmpty ())
     defaultFileName = defaultDirectory;
